+++ Tzkt.Api/Services/Cache/State/RawModels/RawState.cs
@@ -38,7 +38,7 @@
         public int DelegationOpsCount { get; set; }
         public int DoubleBakingOpsCount { get; set; }
         public int DoubleConsensusOpsCount { get; set; }
-        public int AttestationOpsCount { get; set; }
+        public long AttestationOpsCount { get; set; }
         public int PreattestationOpsCount { get; set; }
         public int NonceRevelationOpsCount { get; set; }
         public int VdfRevelationOpsCount { get; set; }
