+++ Tzkt.Data/Models/AppState.cs
@@ -59,7 +59,7 @@ namespace Tzkt.Data.Models
         public int DalEntrapmentEvidenceOpsCount { get; set; }
         public int DoubleBakingOpsCount { get; set; }
         public int DoubleConsensusOpsCount { get; set; }
-        public int AttestationOpsCount { get; set; }
+        public long AttestationOpsCount { get; set; }
         public int PreattestationOpsCount { get; set; }
         public int NonceRevelationOpsCount { get; set; }
         public int VdfRevelationOpsCount { get; set; }
